// studio/sanity.types.ts
export type TestimonialSection = {
  _type: 'testimonialSection'
  topPadding?: boolean
  backgroundColor?: SimplerColor
  testimonialTiles?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'quote'
  }>
}
// ...
export type VideoSection = {
  _type: 'videoSection'
  topPadding?: boolean
  video?: VideoId
}
// ...
export type RecipesSection = {
  _type: 'recipesSection'
  backgroundColour?: SimplerColor
  showRecipesArchive?: ShowRecipesArchive
  verticalSpace?: VerticalSpace
}
// ...
export type HeaderSection = {
  _type: 'headerSection'
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  title?: string
  alignment?: TextAlign
  text?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'body2'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  links?: Array<
    {
      _key: string
    } & LinkDefined
  >
  backgroundColour?: SimplerColor
  verticalSpace?: VerticalSpace
}

export type HotspotSection = {
  _type: 'hotspotSection'
  title?: string
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  hotspots?: Array<
    {
      _key: string
    } & HotspotItem
  >
}

export type BorderSection = {
  _type: 'borderSection'
  borderType?: 'Type 2' | 'Type 3' | 'Type 6'
  borderDirection?: 'Top' | 'Bottom'
  mirror?: boolean
  joiningColour?: SimplerColor
  backgroundColour?: SimplerColor
}

export type TextSection = {
  _type: 'textSection'
  text?: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<
          | ({
              _key: string
            } & TextColor)
          | ({
              _key: string
            } & TextColumns)
          | {
              reference?:
                | {
                    _ref: string
                    _type: 'reference'
                    _weak?: boolean
                    [internalGroqTypeReferenceTo]?: 'post'
                  }
                | {
                    _ref: string
                    _type: 'reference'
                    _weak?: boolean
                    [internalGroqTypeReferenceTo]?: 'page'
                  }
                | {
                    _ref: string
                    _type: 'reference'
                    _weak?: boolean
                    [internalGroqTypeReferenceTo]?: 'recipes'
                  }
              _type: 'internalLink'
              _key: string
            }
          | {
              href?: string
              blank?: boolean
              _type: 'link'
              _key: string
            }
          | {
              asset?: {
                _ref: string
                _type: 'reference'
                _weak?: boolean
                [internalGroqTypeReferenceTo]?: 'sanity.fileAsset'
              }
              _type: 'file'
              _key: string
            }
          | ({
              _key: string
            } & TextColor)
        >
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        _type: 'image'
        _key: string
      }
  >
  backgroundColour?: SimplerColor
  link?: LinkDefined
  verticalSpace?: VerticalSpace
  sideAssets?: SupportingAssets
}

export type TitleSection = {
  _type: 'titleSection'
  title?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5'
    listItem?: never
    markDefs?: Array<
      {
        _key: string
      } & TextColor
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  backgroundColour?: SimplerColor
  link?: LinkDefined
  verticalSpace?: VerticalSpace
  alignment?: TextAlign
  sideAssets?: SupportingAssets
}

export type LinkDefined = {
  _type: 'linkDefined'
  text?: string
  link?: Link
}

export type EmbedSection = {
  _type: 'embedSection'
  scriptTag?: string
  embedCode?: Code
  backgroundColour?: SimplerColor
  verticalSpace?: VerticalSpace
  containerWidth?: ContainerWidth
}

export type FeatureSection = {
  _type: 'featureSection'
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  text?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'overline'
    listItem?: 'bullet' | 'number'
    markDefs?: Array<
      | ({
          _key: string
        } & TextColor)
      | ({
          _key: string
        } & TextColumns)
      | {
          reference?:
            | {
                _ref: string
                _type: 'reference'
                _weak?: boolean
                [internalGroqTypeReferenceTo]?: 'post'
              }
            | {
                _ref: string
                _type: 'reference'
                _weak?: boolean
                [internalGroqTypeReferenceTo]?: 'page'
              }
            | {
                _ref: string
                _type: 'reference'
                _weak?: boolean
                [internalGroqTypeReferenceTo]?: 'recipes'
              }
          _type: 'internalLink'
          _key: string
        }
      | {
          href?: string
          blank?: boolean
          _type: 'link'
          _key: string
        }
      | {
          asset?: {
            _ref: string
            _type: 'reference'
            _weak?: boolean
            [internalGroqTypeReferenceTo]?: 'sanity.fileAsset'
          }
          _type: 'file'
          _key: string
        }
      | ({
          _key: string
        } & TextColor)
    >
    level?: number
    _type: 'block'
    _key: string
  }>
  links?: Array<
    {
      _key: string
    } & LinkDefined
  >
  mirror?: boolean
  backgroundColour?: SimplerColor
  centerAsset?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  verticalSpace?: VerticalSpace
}
// ...
export type ContainerWidth = {
  _type: 'containerWidth'
  width?: 'sm' | 'md' | 'lg'
}
// ...
export type Post = {
  _id: string
  _type: 'post'
  _createdAt: string
  _updatedAt: string
  _rev: string
  metaTitle?: string
  metaDescription?: string
  ogTitle?: string
  ogDescription?: string
  ogImage?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  locale?: 'en-US' | 'en-GB' | 'fr-FR'
  title?: string
  slug?: Slug
  date?: string
  modified?: string
  status?:
    | 'publish'
    | 'future'
    | 'draft'
    | 'pending'
    | 'private'
    | 'trash'
    | 'auto-draft'
    | 'inherit'
  pageBuilder?: Array<
    | ({
        _key: string
      } & EmbedSection)
    | ({
        _key: string
      } & TitleSection)
    | ({
        _key: string
      } & TextSection)
    | ({
        _key: string
      } & FeatureSection)
    | ({
        _key: string
      } & BorderSection)
    | ({
        _key: string
      } & HotspotSection)
    | ({
        _key: string
      } & HeaderSection)
    | ({
        _key: string
      } & TestimonialSection)
    | ({
        _key: string
      } & TeamSection)
    | ({
        _key: string
      } & RecipesSection)
    | ({
        _key: string
      } & VideoSection)
    | ({
        _key: string
      } & FeaturesListSection)
    | ({
        _key: string
      } & CtaSection)
    | ({
        _key: string
      } & NewsletterSection)
    | ({
        _key: string
      } & ImageCarouselSection)
    | ({
        _key: string
      } & BlogSection)
    | ({
        _key: string
      } & StepsSection)
    | ({
        _key: string
      } & TimelineSection)
    | ({
        _key: string
      } & LocationSection)
    | ({
        _key: string
      } & ContactSection)
    | ({
        _key: string
      } & ClientLoginSection)
  >
  content?: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        _type: 'image'
        _key: string
      }
  >
  excerpt?: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
        listItem?: 'bullet' | 'number'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        _type: 'image'
        _key: string
      }
  >
  featuredMedia?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  tileImage?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  author?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'author'
  }
  categories?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'categories'
  }>
}
// ...
export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export type SimplerColor = {
  _type: 'simplerColor'
  label?: string
  value?: string
}

export type Code = {
  _type: 'code'
  language?: string
  filename?: string
  code?: string
  highlightedLines?: Array<number>
}
// ...
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol